+++ app/services/candidate_interface/application_not_ready_to_send_error.rb
@@ -1,0 +1,13 @@
+module CandidateInterface
+  class ApplicationNotReadyToSendError < StandardError
+    attr_reader :application_choice
+
+    def initialize(application_choice)
+      @application_choice = application_choice
+    end
+
+    def message
+      "Tried to send an application in the #{application_choice.status} state to a provider"
+    end
+  end
+end
